chore(`cheatcodes`): clean up cheatcodes `util.rs` with new `wallet` and `parsing` (#5758)
* feat: add new wallet/parsing files with respective funcs

* chore: dedupe

* chore: re-use new cheatcodes

* chore: move skip to ext

* fmt

* chore: parsing -> parse
