+++ crates/evm/src/executor/inspector/cheatcodes/ext.rs
@@ -1,5 +1,5 @@
-use super::{bail, ensure, fmt_err, Cheatcodes, Result};
-use crate::{abi::HEVMCalls, executor::inspector::cheatcodes::util};
+use super::{bail, ensure, fmt_err, util::MAGIC_SKIP_BYTES, Cheatcodes, Error, Result};
+use crate::{abi::HEVMCalls, executor::inspector::cheatcodes::parse};
 use ethers::{
     abi::{self, AbiEncode, JsonAbi, ParamType, Token},
     prelude::artifacts::CompactContractBytecode,
@@ -7,6 +7,7 @@ use ethers::{
 };
 use foundry_common::{fmt::*, fs, get_artifact_path};
 use foundry_config::fs_permissions::FsAccessKind;
+use revm::{Database, EVMData};
 use serde::Deserialize;
 use serde_json::Value;
 use std::{collections::BTreeMap, env, path::Path, process::Command};
@@ -190,9 +191,9 @@ fn get_env(key: &str, ty: ParamType, delim: Option<&str>, default: Option<String
         })
     })?;
     if let Some(d) = delim {
-        util::parse_array(val.split(d).map(str::trim), &ty)
+        parse::parse_array(val.split(d).map(str::trim), &ty)
     } else {
-        util::parse(&val, &ty)
+        parse::parse(&val, &ty)
     }
 }
 
@@ -342,9 +343,9 @@ fn parse_json(json_str: &str, key: &str, coerce: Option<ParamType>) -> Result {
                 };
                 trace!(target : "forge::evm", ?values, "parsign values");
                 return if let Some(array) = values[0].as_array() {
-                    util::parse_array(array.iter().map(to_string), &coercion_type)
+                    parse::parse_array(array.iter().map(to_string), &coercion_type)
                 } else {
-                    util::parse(&to_string(values[0]), &coercion_type)
+                    parse::parse(&to_string(values[0]), &coercion_type)
                 }
             }
 
@@ -482,7 +483,7 @@ fn key_exists(json_str: &str, key: &str) -> Result {
     let json: Value =
         serde_json::from_str(json_str).map_err(|e| format!("Could not convert to JSON: {e}"))?;
     let values = jsonpath_lib::select(&json, &canonicalize_json_key(key))?;
-    let exists = util::parse(&(!values.is_empty()).to_string(), &ParamType::Bool)?;
+    let exists = parse::parse(&(!values.is_empty()).to_string(), &ParamType::Bool)?;
     Ok(exists)
 }
 
@@ -494,8 +495,28 @@ fn sleep(milliseconds: &U256) -> Result {
     Ok(Default::default())
 }
 
+/// Skip the current test, by returning a magic value that will be checked by the test runner.
+pub fn skip(state: &mut Cheatcodes, depth: u64, skip: bool) -> Result {
+    if !skip {
+        return Ok(b"".into())
+    }
+
+    // Skip should not work if called deeper than at test level.
+    // As we're not returning the magic skip bytes, this will cause a test failure.
+    if depth > 1 {
+        return Err(Error::custom("The skip cheatcode can only be used at test level"))
+    }
+
+    state.skip = true;
+    Err(Error::custom_bytes(MAGIC_SKIP_BYTES))
+}
+
 #[instrument(level = "error", name = "ext", target = "evm::cheatcodes", skip_all)]
-pub fn apply(state: &mut Cheatcodes, call: &HEVMCalls) -> Option<Result> {
+pub fn apply<DB: Database>(
+    state: &mut Cheatcodes,
+    data: &mut EVMData<'_, DB>,
+    call: &HEVMCalls,
+) -> Option<Result> {
     Some(match call {
         HEVMCalls::Ffi(inner) => {
             if state.config.ffi {
@@ -680,6 +701,7 @@ pub fn apply(state: &mut Cheatcodes, call: &HEVMCalls) -> Option<Result> {
         HEVMCalls::WriteJson0(inner) => write_json(state, &inner.0, &inner.1, None),
         HEVMCalls::WriteJson1(inner) => write_json(state, &inner.0, &inner.1, Some(&inner.2)),
         HEVMCalls::KeyExists(inner) => key_exists(&inner.0, &inner.1),
+        HEVMCalls::Skip(inner) => skip(state, data.journaled_state.depth(), inner.0),
         _ => return None,
     })
 }
