+++ crates/evm/src/executor/inspector/cheatcodes/mod.rs
@@ -57,10 +57,15 @@ mod fs;
 mod fuzz;
 /// Mapping related cheatcodes
 mod mapping;
+/// Parsing related cheatcodes.
+/// Does not include JSON-related cheatcodes to cut complexity.
+mod parse;
 /// Snapshot related cheatcodes
 mod snapshot;
-/// Utility cheatcodes (`sign` etc.)
+/// Utility functions and constants.
 pub mod util;
+/// Wallet / key management related cheatcodes
+mod wallet;
 pub use util::{BroadcastableTransaction, DEFAULT_CREATE2_DEPLOYER};
 
 mod config;
@@ -219,10 +224,11 @@ impl Cheatcodes {
 
         let opt = env::apply(self, data, caller, &decoded)
             .transpose()
-            .or_else(|| util::apply(self, data, &decoded))
+            .or_else(|| wallet::apply(self, data, &decoded))
+            .or_else(|| parse::apply(self, data, &decoded))
             .or_else(|| expect::apply(self, data, &decoded))
             .or_else(|| fuzz::apply(&decoded))
-            .or_else(|| ext::apply(self, &decoded))
+            .or_else(|| ext::apply(self, data, &decoded))
             .or_else(|| fs::apply(self, &decoded))
             .or_else(|| snapshot::apply(data, &decoded))
             .or_else(|| fork::apply(self, data, &decoded));
